chore(tech): extract stripEmptyFormFields helper into it's own directory
