+++ src/ui/server/helpers/strip-empty-form-fields/index.ts
@@ -1,0 +1,31 @@
+import { RequestBody } from '../../../types';
+
+/**
+ * stripEmptyFormFields
+ * Filters out form fields with empty values
+ * @param {Express.Request.body} Form data
+ * @returns {Object} Form data without empty values
+ */
+const stripEmptyFormFields = (formBody: RequestBody) => {
+  if (formBody) {
+    const fieldsWithValues = {};
+
+    const keys = Object.keys(formBody);
+
+    keys.forEach((key) => {
+      const value = formBody[key];
+
+      const hasValue = Boolean(value) || value === false;
+
+      if (hasValue) {
+        fieldsWithValues[key] = value;
+      }
+    });
+
+    return fieldsWithValues;
+  }
+
+  return {};
+};
+
+export default stripEmptyFormFields;
